Fix int-to-size_t warnings

+++ libCloudundancyTests/ValueTypes/FilePathLineNumberLineTextTests.cpp
@@ -36,7 +36,7 @@ TEST(Equalizer_ThrowsIfAnyFieldNotEqual)
 {
    ZENUNIT_EQUALIZER_TEST_SETUP(FilePathLineNumberLineText);
    ZENUNIT_EQUALIZER_THROWS_WHEN_FIELD_NOT_EQUAL(FilePathLineNumberLineText, filePath, ZenUnit::Random<fs::path>());
-   ZENUNIT_EQUALIZER_THROWS_WHEN_FIELD_NOT_EQUAL(FilePathLineNumberLineText, lineNumber, ZenUnit::RandomNon0<int>());
+   ZENUNIT_EQUALIZER_THROWS_WHEN_FIELD_NOT_EQUAL(FilePathLineNumberLineText, lineNumber, ZenUnit::RandomNon0<size_t>());
    ZENUNIT_EQUALIZER_THROWS_WHEN_FIELD_NOT_EQUAL(FilePathLineNumberLineText, lineText, ZenUnit::Random<string>());
 }
